Restructure to not have it all in one file
