+++ src/api/plugin_instruction_handler.rs
@@ -1,5 +1,7 @@
-use super::schema::api::*;
-
+use super::schema::{
+    auth::AuthAccountInstruction,
+    keepalive::KeepaliveInstruction,
+};
 /// A trait to be implemented by the core for instructions sent from a plugin
 /// Also can be implemented by the plugin SDK, which can then be translated
 /// to instructions, and back again in the core.
