+++ src/api/schema/mod.rs
@@ -1,1 +1,4 @@
-pub mod api;
+pub mod auth;
+pub mod instructions;
+pub mod keepalive;
+pub mod protocol;
